backend/requirements.txt
flask==3.0.0
flask-cors==4.0.0
numpy==1.26.4
opencv-python==4.8.1.78
scipy==1.11.4
torch==2.0.1
torchvision==0.15.2
pillow==10.1.0
scikit-learn==1.3.2
